fix try

+++ stargateSharded.php
@@ -344,22 +344,25 @@ $discord->on('ready', function ($discord) use($beta){
                 $reminder->delete();
             else
             {
-                $userExist = $discord->factory(\Discord\Parts\User\User::class, [
-                    'id' => $reminder->player->user_id,
-                ]);
-                //$userExist = $discord->users->get('id',$reminder->player->user_id);
-                if(!is_null($userExist))
-                {
-                    if(!is_null($reminder->embed))
+                $discord->users->fetch($reminder->player->user_id)->done(function(User $userExist) use($reminder,$discord){
+                    if(!is_null($userExist))
                     {
-                        $reminderEmbed = json_decode($reminder->embed,true);
-                        $newEmbed = $discord->factory(Embed::class,$reminderEmbed);
-                        $userExist->sendMessage('', false, $newEmbed);
+                        if(!is_null($reminder->embed))
+                        {
+                            $reminderEmbed = json_decode($reminder->embed,true);
+                            $newEmbed = $discord->factory(Embed::class,$reminderEmbed);
+                            $userExist->sendMessage('', false, $newEmbed)->done(function(Message $message) use($reminder){
+                                if(!is_null($message))
+                                    $reminder->delete();
+                            });
+                        }
+                        else
+                            $userExist->sendMessage($reminder->reminder)->done(function(Message $message) use($reminder){
+                                if(!is_null($message))
+                                    $reminder->delete();
+                            });
                     }
-                    else
-                        $userExist->sendMessage($reminder->reminder);
-                    $reminder->delete();
-                }
+                });
             }
         }
     });
